refactor: Controller 로직 재분배
